Implement converter from filters to safari blocking rules
