+++ src/matching/cosmetics.ts
@@ -1,7 +1,7 @@
 // @ts-ignore
 import { getPublicSuffix } from 'tldjs';
 
-import { CosmeticFilter } from '../parsing/cosmetic-filter';
+import CosmeticFilter from '../types/cosmetics';
 import { fastStartsWith } from '../utils';
 
 /* Checks that hostnamePattern matches at the end of the hostname.
