+++ src/parsing/cosmetic-filter.ts
@@ -4,157 +4,9 @@ import {
   fastStartsWithFrom,
   getBit,
   setBit,
-  tokenizeCSS,
 } from '../utils';
-import IFilter from './interface';
 
-/**
- * Masks used to store options of cosmetic filters in a bitmask.
- */
-const enum COSMETICS_MASK {
-  unhide = 1 << 0,
-  scriptInject = 1 << 1,
-  scriptBlock = 1 << 2,
-}
-
-/***************************************************************************
- *  Cosmetic filters parsing
- * ************************************************************************ */
-
-export class CosmeticFilter implements IFilter {
-  public id: number;
-  public mask: number;
-  public selector: string;
-  public hostnames: string;
-
-  // For debug only
-  public rawLine: string | null;
-
-  private hostnamesArray: string[] | null;
-
-  constructor({
-    mask,
-    selector,
-    hostnames,
-    id,
-  }: {
-    mask: number,
-    selector: string,
-    hostnames: string,
-    id: number,
-  }) {
-    this.id = id;
-    this.mask = mask;
-    this.selector = selector;
-    this.hostnames = hostnames;
-
-    // Lazily set when needed
-    this.hostnamesArray = null;
-
-    // Only in debug mode
-    this.rawLine = null;
-  }
-
-  public isCosmeticFilter(): boolean {
-    return true;
-  }
-  public isNetworkFilter(): boolean {
-    return false;
-  }
-
-  /**
-   * Create a more human-readable version of this filter. It is mainly used for
-   * debugging purpose, as it will expand the values stored in the bit mask.
-   */
-  public toString(): string {
-    let filter = '';
-
-    if (this.hasHostnames()) {
-      filter += this.hostnames;
-    }
-
-    if (this.isUnhide()) {
-      filter += '#@#';
-    } else {
-      filter += '##';
-    }
-
-    if (this.isScriptInject()) {
-      filter += 'script:inject(';
-      filter += this.selector;
-      filter += ')';
-    } else if (this.isScriptBlock()) {
-      filter += 'script:contains(';
-      filter += this.selector;
-      filter += ')';
-    } else {
-      filter += this.selector;
-    }
-
-    return filter;
-  }
-
-  public getTokens(): number[][] {
-    return [this.getTokensSelector()];
-  }
-
-  public getTokensSelector(): number[] {
-    if (this.isScriptInject() || this.isScriptBlock()) {
-      return [];
-    }
-
-    // Only keep the part after the last '>'
-    const sepIndex = this.selector.lastIndexOf('>');
-    if (sepIndex !== -1) {
-      return tokenizeCSS(this.selector.substr(sepIndex));
-    }
-
-    return tokenizeCSS(this.selector);
-  }
-
-  public getSelector(): string {
-    return this.selector;
-  }
-
-  public hasHostnames(): boolean {
-    return !!this.hostnames;
-  }
-
-  public getHostnames(): string[] {
-    if (this.hostnamesArray === null) {
-      if (this.hasHostnames()) {
-        // Sort them from longer hostname to shorter.
-        // This is to make sure that we will always start by the most specific
-        // when matching.
-        this.hostnamesArray = this.hostnames.split(',').sort((h1, h2) => {
-          if (h1.length > h2.length) {
-            return -1;
-          } else if (h1.length < h2.length) {
-            return 1;
-          }
-
-          return 0;
-        });
-      } else {
-        this.hostnamesArray = [];
-      }
-    }
-
-    return this.hostnamesArray;
-  }
-
-  public isUnhide(): boolean {
-    return getBit(this.mask, COSMETICS_MASK.unhide);
-  }
-
-  public isScriptInject(): boolean {
-    return getBit(this.mask, COSMETICS_MASK.scriptInject);
-  }
-
-  public isScriptBlock(): boolean {
-    return getBit(this.mask, COSMETICS_MASK.scriptBlock);
-  }
-}
+import CosmeticFilter, { MASK } from '../types/cosmetics';
 
 /**
  * Given a line that we know contains a cosmetic filter, create a CosmeticFiler
@@ -164,8 +16,8 @@ export class CosmeticFilter implements IFilter {
 export function parseCosmeticFilter(line: string): CosmeticFilter | null {
   // Mask to store attributes
   // Each flag (unhide, scriptInject, etc.) takes only 1 bit
-  // at a specific offset defined in COSMETICS_MASK.
-  // cf: COSMETICS_MASK for the offset of each property
+  // at a specific offset defined in MASK.
+  // cf: MASK for the offset of each property
   let mask = 0;
   let selector: string = '';
   let hostnames: string = ''; // Coma-separated list of hostnames
@@ -184,7 +36,7 @@ export function parseCosmeticFilter(line: string): CosmeticFilter | null {
 
   // Check if unhide
   if (line[afterSharpIndex] === '@') {
-    mask = setBit(mask, COSMETICS_MASK.unhide);
+    mask = setBit(mask, MASK.unhide);
     suffixStartIndex += 1;
   }
 
@@ -216,10 +68,10 @@ export function parseCosmeticFilter(line: string): CosmeticFilter | null {
     let scriptSelectorIndexEnd = selector.length - 1;
 
     if (fastStartsWithFrom(selector, 'inject(', scriptMethodIndex)) {
-      mask = setBit(mask, COSMETICS_MASK.scriptInject);
+      mask = setBit(mask, MASK.scriptInject);
       scriptSelectorIndexStart += 'inject('.length;
     } else if (fastStartsWithFrom(selector, 'contains(', scriptMethodIndex)) {
-      mask = setBit(mask, COSMETICS_MASK.scriptBlock);
+      mask = setBit(mask, MASK.scriptBlock);
       scriptSelectorIndexStart += 'contains('.length;
 
       // If it's a regex
@@ -244,7 +96,7 @@ export function parseCosmeticFilter(line: string): CosmeticFilter | null {
     selector.length === 0 ||
     selector.endsWith('}') ||
     selector.indexOf('##') !== -1 ||
-    (getBit(mask, COSMETICS_MASK.unhide) && hostnames.length === 0)
+    (getBit(mask, MASK.unhide) && hostnames.length === 0)
   ) {
     return null;
   }
