+++ tsconfig.json
@@ -22,15 +22,16 @@
     "build"
   ],
   "files": [
-    "./index.ts",
+    "./assets/load.js",
     "./index-cosmetics.ts",
     "./index-matching.ts",
-    "test/engine.test.ts",
-    "test/injection.test.ts",
-    "test/matching.test.js",
-    "test/parsing.test.ts",
-    "test/serialization.test.ts",
-    "test/utils.js",
-    "test/utils.test.ts"
+    "./index.ts",
+    "./scripts/convert-to-safari-rules.ts",
+    "./test/engine.test.ts",
+    "./test/injection.test.ts",
+    "./test/matching.test.js",
+    "./test/parsing.test.ts",
+    "./test/serialization.test.ts",
+    "./test/utils.test.ts"
   ]
 }
